pyproject.toml
[tool.poetry]
name = "task2week_cli_sql"
version = "0.1.0"
description = ""
authors = ["satohitoshi1 <satohitoshi121@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.20.0"
peewee = "^3.15.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
